+++ packages/jsonapi/src/Validation/FieldsException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EDT\JsonApi\Validation;
+
+use InvalidArgumentException;
+
+class FieldsException extends InvalidArgumentException
+{
+
+}